config.py
import os
import json
from dotenv import load_dotenv

load_dotenv()


class Config:
    TELEGRAM_TOKEN = os.getenv("TELEGRAM_TOKEN")
    SHEET_ID = os.getenv("SHEET_ID")
    REDIS_URL = os.getenv("REDIS_URL", "redis://redis:6379/0")
    LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
    SESSION_TTL_PADDING = int(os.getenv("SESSION_TTL_PADDING", "300"))

    # Google Sheets credentials
    _credentials_value = os.getenv("GOOGLE_CREDENTIALS")
    if _credentials_value:
        try:
            GOOGLE_CREDENTIALS = json.loads(_credentials_value)
        except json.JSONDecodeError:
            # Если это путь к файлу
            credentials_path = os.path.abspath(_credentials_value)
            if os.path.exists(credentials_path):
                with open(credentials_path, "r", encoding="utf-8") as f:
                    GOOGLE_CREDENTIALS = json.load(f)
            else:
                raise ValueError(
                    "GOOGLE_CREDENTIALS должен быть валидным JSON или путем к файлу"
                )
    else:
        raise ValueError("GOOGLE_CREDENTIALS не установлен")

    @classmethod
    def validate(cls):
        if not cls.TELEGRAM_TOKEN:
            raise ValueError("TELEGRAM_TOKEN не установлен")
        if not cls.SHEET_ID:
            raise ValueError("SHEET_ID не установлен")

