requirements.txt
aiogram==3.13.1
redis==5.2.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.152.0
python-dotenv==1.0.1
pytz==2024.2

